+++ src/main/java/com/cgl/model/CustomerType.java
@@ -13,19 +13,21 @@ import javax.persistence.Table;
 @Entity
 @Table(name = "t_CustomerType")
 public class CustomerType {
-    private int ct_id;//主键
+    private long ct_id;//主键
     private  String ct_name;//客户分类，名
+
     
+
     @Id
     @GeneratedValue
-    public int getCt_id() {
+    public long getCt_id() {
         return ct_id;
     }
 
-    public void setCt_id(int ct_id) {
+    public void setCt_id(long ct_id) {
         this.ct_id = ct_id;
     }
-
+    
     public String getCt_name() {
         return ct_name;
     }
